chore: Move upgrade logic to utility
