+++ src/Connector.SqlServer/Connector/SqlServerConnector.cs
@@ -1,6 +1,7 @@
 ﻿using CluedIn.Connector.SqlServer.Exceptions;
 using CluedIn.Connector.SqlServer.Utils;
 using CluedIn.Connector.SqlServer.Utils.TableDefinitions;
+using CluedIn.Connector.SqlServer.Utils.Upgrade;
 using CluedIn.Core;
 using CluedIn.Core.Connectors;
 using CluedIn.Core.Data.Parts;
@@ -78,84 +79,7 @@ namespace CluedIn.Connector.SqlServer.Connector
 
                 var mainTableName = TableNameUtility.GetMainTableName(streamModel, schema);
 
-                // Archive old edge table name, since rename logic used during normal archive expect different name
-                {
-                    var oldEdgeLocalTableName = $"{mainTableName.LocalName}Edges";
-                    var newEdgeLocalTableName = $"{mainTableName.LocalName}_{DateTimeOffset.UtcNow}Edges";
-
-                    var oldEdgeTableName = SqlName.FromUnsafe(oldEdgeLocalTableName).ToTableName(schema);
-                    var newEdgeTableName = SqlName.FromUnsafe(newEdgeLocalTableName).ToTableName(schema);
-
-                    var tableRenameText = $"""
-                        IF (OBJECT_ID(N'{oldEdgeTableName.FullyQualifiedName}') IS NOT NULL)
-                        BEGIN
-                            EXEC sp_rename N'{oldEdgeTableName.FullyQualifiedName}', {newEdgeTableName.LocalName};
-                        END
-                        """;
-
-                    var renameSqlConnectorCommand = new SqlServerConnectorCommand() { Text = tableRenameText, Parameters = Array.Empty<SqlParameter>() };
-                    await renameSqlConnectorCommand
-                        .ToSqlCommand(transaction)
-                        .ExecuteScalarAsync();
-                }
-
-                // Check if old main table is present
-                {
-                    var tableExistsText = $"""
-                        IF (OBJECT_ID(N'{mainTableName.FullyQualifiedName}') IS NOT NULL)
-                        BEGIN
-                        	SELECT 1
-                        END
-                        ELSE
-                        BEGIN
-                        	SELECT 0
-                        END
-                        """;
-
-                    var tableExistsCommand = new SqlServerConnectorCommand() { Text = tableExistsText, Parameters = Array.Empty<SqlParameter>() };
-                    var tableExistsResult = await tableExistsCommand.ToSqlCommand(transaction).ExecuteScalarAsync();
-
-                    // If the table exists, we need to check if it was created in an older version of the connector.
-                    // The way we do this is check is to check if it contains all of the columns that we expect the
-                    // main table to have in this version.
-                    if ((int)tableExistsResult == 1)
-                    {
-                        var tableColumnsSelectText = $"""
-                        SELECT columns.name FROM sys.columns columns
-                        INNER JOIN sys.tables tables
-                        ON tables.object_id = columns.object_id AND 
-                           tables.Name = '{mainTableName.LocalName}' AND
-                           tables.type = 'U'
-                        """;
-
-                        var tableCheckSqlConnectorCommand = new SqlServerConnectorCommand() { Text = tableColumnsSelectText, Parameters = Array.Empty<SqlParameter>() };
-                        var reader = await tableCheckSqlConnectorCommand
-                            .ToSqlCommand(transaction)
-                            .ExecuteReaderAsync();
-
-                        var existingColumns = new List<string>();
-                        while (await reader.ReadAsync())
-                        {
-                            existingColumns.Add(reader[0].ToString());
-                        }
-
-                        var expectedColumnNames = MainTableDefinition
-                            .GetColumnDefinitions(StreamMode.Sync, Array.Empty<(string, ConnectorPropertyDataType)>())
-                            .Select(columnDefinition => columnDefinition.Name);
-
-                        var existingColumnsContainsAllExpectedColumns = expectedColumnNames.All(column => existingColumns.Contains(column));
-
-                        if (!existingColumnsContainsAllExpectedColumns)
-                        {
-                            // If an exception is thrown during `VerifyExistingContainer`, nothing can be done with the stream.
-                            // This includes reprocessing the stream, to create new tables.
-                            // Until this is changed in platform, we simply log the exception instead of throwing
-                            // PBI: #23500
-                            var exception = IncompatibleTableException.OldTableVersionExists(streamModel.Id, (Guid)streamModel.ConnectorProviderDefinitionId);
-                            _logger.LogError(exception, "Not all expected columns were present, most likely because the table was created in an old version");
-                        }
-                    }
-                }
+                await UpgradeTo370Utility.Upgrade(streamModel, mainTableName, schema, transaction, _logger);
 
                 await transaction.CommitAsync();
             });
